scheduler : new feature of command schedule processing
Number of attempts and delays between failed attempts are automaticaly calculated and no longer accessibles for user.
This values are calculated and based on statical analysis during the valid period.

This process can be impacted by new parameter "web_def_attempts_per_day" - average of number of attempts per day.
Number of attempts is more frequently on the start and end of valid period.